refactor: move fixtures to separate file
to avoid reinit
