+++ setup.cfg
@@ -35,6 +35,10 @@ ignore_missing_imports = True
 ignore_errors = True
 
 [tool:pytest]
+# -p no:cacheprovider
+addopts =
+    # *.xsh specific tests need to be run separately to avoid double initialization of XSH
+    -p no:xonsh
 cache_dir = .cache/pytest
 markers =
     news: check changelog unit is valid rst
