+++ tests/conftest.py
@@ -1,2 +1,2 @@
 # when xonsh is not installed in the current env
-# pytest_plugins = ("xonsh.pytest.plugin",)
+pytest_plugins = ("xonsh.pytest.fixtures",)
